+++ mamr WatchKit Extension/ExtensionDelegate.swift
@@ -31,6 +31,7 @@ class ExtensionDelegate: NSObject, WKExtensionDelegate, UNUserNotificationCenter
         }
     }
     
+    
     func userNotificationCenter(_ center: UNUserNotificationCenter, didReceive response: UNNotificationResponse, withCompletionHandler completionHandler: @escaping () -> Void) {
         print("Tapped in notification")
         let identifier = response.actionIdentifier
@@ -45,10 +46,7 @@ class ExtensionDelegate: NSObject, WKExtensionDelegate, UNUserNotificationCenter
         }
         completionHandler()
     }
-   
-    func applicationDidFinishLaunchingwithoptions() {
-    }
-    
+
 
     func applicationDidBecomeActive() {
         // Restart any tasks that were paused (or not yet started) while the application was inactive. If the application was previously in the background, optionally refresh the user interface.
