+++ mamr WatchKit Extension/InterfaceController.swift
@@ -14,31 +14,13 @@ import HealthKit
 
 class InterfaceController: WKInterfaceController, UNUserNotificationCenterDelegate { //Extending class with notifications delegate
     
-    func lunchnotificationcategories() { //declaring for actionable buttons for lunchnotifications
-        
-        let order1 = UNNotificationAction(identifier: "order1", title: "Order Dish #1", options: .foreground) //Button 1
-        let order2 = UNNotificationAction(identifier: "order2", title: "Order Dish #2", options: .foreground) //Button 2
-        
-        let lunchcategory = UNNotificationCategory(identifier: "lunch_notification", actions: [order1, order2], intentIdentifiers: []) //setting the notification category for lunch
-        
-        UNUserNotificationCenter.current().setNotificationCategories([lunchcategory]) //setting actions & categories
-    }
-    
-    
-    func dinnernotificationcategories() { //declaring for actionable buttons for dinnernotifications
-        
-        let order1 = UNNotificationAction(identifier: "getrecipe1", title: "Get Recipe #1", options: .foreground) //Button 1
-        let order2 = UNNotificationAction(identifier: "getrecipe2", title: "Get Recipe #2", options: .foreground) //Button 2
-        
-        let dinnercategory = UNNotificationCategory(identifier: "dinner_notification", actions: [order1, order2], intentIdentifiers: []) //setting the notification category for dinner
-        
-        UNUserNotificationCenter.current().setNotificationCategories([dinnercategory])  //setting actions & categories
-    }
-    
+
     
     @IBAction func lunch_notifications() { //lunch sample notification
         
-        lunchnotificationcategories()
+        let notificationManger = Notificationshandler()
+        
+        notificationManger.lunchnotificationcategories()
 
         let center = UNUserNotificationCenter.current()
         center.removeAllPendingNotificationRequests()
@@ -62,7 +44,8 @@ class InterfaceController: WKInterfaceController, UNUserNotificationCenterDelega
     
     @IBAction func dinner_notifications() { //dinner sample notification
         
-        dinnernotificationcategories()
+        let notificationManger = Notificationshandler()
+        notificationManger.dinnernotificationcategories()
         
         let center = UNUserNotificationCenter.current()
         center.removeAllPendingNotificationRequests()
